+++ test/type/mutator.ts
@@ -117,58 +117,3 @@ export function useConfigMutate() {
     return '0'
   })
 }
-
-export function useTestSWRMutation() {
-  const { data } = useSWR('key', async () => {
-    return ['foo']
-  })
-  const { trigger } = useSWRMutation(
-    'key',
-    async (_, { arg }: { arg: 'foo' }) => {
-      return arg.toUpperCase()
-    }
-  )
-
-  const test = () => {
-    // @ts-expect-error `arg` should be 'foo'
-    trigger()
-
-    // @ts-expect-error `arg` should be 'foo'
-    trigger<typeof data>('bar', {
-      optimisticData: current => {
-        expectType<string[] | undefined>(current)
-        return []
-      },
-      populateCache: (added, current) => {
-        expectType<string>(added)
-        expectType<typeof data>(current)
-        return []
-      },
-      revalidate: false
-    })
-  }
-  test()
-}
-
-export function useTestSWRMutationWithSWRMutate() {
-  const { mutate } = useSWR('/some/key', () => {
-    return {
-      foo: 'bar'
-    }
-  })
-  const { trigger } = useSWRMutation('/some/key', () => {
-    return {
-      foo: 'foo'
-    }
-  })
-  const test = () => {
-    ;async () => {
-      mutate(trigger(), {
-        optimisticData: {
-          foo: 'baz'
-        }
-      })
-    }
-  }
-  test()
-}
