+++ test/type/trigger.ts
@@ -1,5 +1,5 @@
 import useSWRMutation from 'swr/mutation'
-
+import useSWR from 'swr'
 type ExpectType = <T>(value: T) => void
 const expectType: ExpectType = () => {}
 
@@ -50,4 +50,59 @@ export function useTriggerWithParameter() {
   // The argument of `trigger` should be number or undefined.
   expectType<Equal<Parameters<typeof trigger>[0], number>>(true)
   expectType<Promise<string | undefined>>(trigger(1))
+}
+
+export function useTestSWRMutation() {
+  const { data } = useSWR('key', async () => {
+    return ['foo']
+  })
+  const { trigger } = useSWRMutation(
+    'key',
+    async (_, { arg }: { arg: 'foo' }) => {
+      return arg.toUpperCase()
+    }
+  )
+
+  const test = () => {
+    // @ts-expect-error `arg` should be 'foo'
+    trigger()
+
+    // @ts-expect-error `arg` should be 'foo'
+    trigger<typeof data>('bar', {
+      optimisticData: current => {
+        expectType<string[] | undefined>(current)
+        return []
+      },
+      populateCache: (added, current) => {
+        expectType<string>(added)
+        expectType<typeof data>(current)
+        return []
+      },
+      revalidate: false
+    })
+  }
+  test()
+}
+
+export function useTestSWRMutationWithSWRMutate() {
+  const { mutate } = useSWR('/some/key', () => {
+    return {
+      foo: 'bar'
+    }
+  })
+  const { trigger } = useSWRMutation('/some/key', () => {
+    return {
+      foo: 'foo'
+    }
+  })
+  const test = () => {
+    ;async () => {
+      mutate(trigger(), {
+        optimisticData: {
+          foo: 'baz'
+        }
+      })
+    }
+  }
+  test()
 }
